choir: separate weights of sub and misleading param of zombienet
