+++ pubsub-dev/zombienet.toml
@@ -32,13 +32,11 @@ cumulus_based = true
     [[parachains.collators]]
     name = "para-collator01"
     rpc_port = 9920
-    chain = "penpal-local"
     command = "./target/release/polkadot-parachain"
     args = ["-lxcm=trace", "--trie-cache-size=0", "--network-backend=libp2p", "--", "--network-backend=libp2p"]
 
     [[parachains.collators]]
     name = "para-collator02"
-    chain = "penpal-local"
     command = "./target/release/polkadot-parachain"
     args = ["-lxcm=trace", "--trie-cache-size=0", "--network-backend=libp2p", "--", "--network-backend=libp2p"]
